Fix notification stream Dart analysis issues
Align the workbench notification subscription field with the void stream produced by asyncMap, update the route overlay fake repository for the new watchConversationEvents contract, and remove the now-unnecessary reconnect delay cast.

Tested: dart format lib\\src\\ui\\features\\workbench\\coding_workbench_page.dart lib\\src\\services\\daemon_notification_client.dart test\\main_route_overlay_test.dart; node scripts/run-tests.js; npm run lint; node scripts/check-project-knowledge.js; git diff --check.

Not-tested: dart run tool\\check_architecture_imports.dart timed out after 121s in this agent environment after the fix; rerun Dart analyze and Flutter tests manually on the user's local shell.

+++ mobile/lib/src/services/daemon_notification_client.dart
@@ -427,7 +427,7 @@ class DaemonNotificationClient implements NotificationService {
     if (reconnectDelays.isEmpty) {
       return _fallbackReconnectDelay;
     }
-    final delayIndex = attempt.clamp(0, reconnectDelays.length - 1) as int;
+    final delayIndex = attempt.clamp(0, reconnectDelays.length - 1);
     return reconnectDelays[delayIndex];
   }
